счета. удаление . невозможно удалить счёт на которой есть фин цель
git-svn-id: svn://hm.podzone.net/trunk@2050 240b7755-b59a-4df0-b20f-7dab6991996d

+++ models/accounts.model.php
@@ -315,8 +315,10 @@ class Accounts_Model
     {
         $fin = new Targets_Model();
         $noFinTarget = $fin->countTargetsOnAccount($id);// если ноль значит удаляем
+        if ($noFinTarget > 0)
+            return 'cel';
         if (!$noFinTarget){
-
+            //return 'del';
             $sql = "DELETE FROM
                         account_field_values
                     WHERE
@@ -331,8 +333,6 @@ class Accounts_Model
             Core::getInstance()->user->save();
             return true;
         }
-        else if ($noFinTarget)
-            return 'cel';
     }
     /**
      * Функция которая отсылает список счетов
